frontend/requirements.txt
streamlit>=1.45.1,<2.0.0
python-dotenv>=1.0.0,<2.0.0
st-cookies-manager>=0.2.2,<0.3.0
streamlit-cookies-controller>=0.0.4,<0.0.5